library_server/categories/serializer.py
from rest_framework import serializers
from .models import Category

class CategorySerializer(serializers.ModelSerializer):
    class Meta:
        model = Category
        fields = ['category_name', 'id']
        
    def __str__(self):
        return super().__str__()