requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
gitdb==4.0.12
GitPython==3.1.41
pip-autoremove==0.10.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
setuptools==80.9.0
smmap==5.0.2
sqlparse==0.5.3
